[Move] fileService from `src/code/common` to `src/code/platform`
