+++ src/code/electron/main.ts
@@ -5,9 +5,9 @@ import { Event } from 'src/base/common/event';
 import { Schemas, URI } from 'src/base/common/file/uri';
 import { BufferLogger, ILogService, LogLevel, PipelineLogger } from 'src/base/common/logger';
 import { Strings } from 'src/base/common/util/string';
-import { DiskFileSystemProvider } from 'src/base/node/diskFileSystemProvider';
+import { DiskFileSystemProvider } from 'src/code/platform/files/node/diskFileSystemProvider';
 import { GlobalConfigService, IGlobalConfigService, IUserConfigService, UserConfigService } from 'src/code/platform/configuration/electron/configService';
-import { FileService, IFileService } from 'src/code/common/service/fileService/fileService';
+import { FileService, IFileService } from 'src/code/platform/files/common/fileService';
 import { IInstantiationService, InstantiationService } from 'src/code/common/service/instantiationService/instantiation';
 import { ServiceCollection } from 'src/code/common/service/instantiationService/serviceCollection';
 import { ILoggerService } from 'src/code/platform/logger/common/abstractLoggerService';
