+++ src/code/platform/files/common/fileService.ts
@@ -6,7 +6,7 @@ import { bufferToStream, IReadableStream, listenStream, newWriteableBufferStream
 import { isAbsoluteURI, URI } from "src/base/common/file/uri";
 import { ILogService } from "src/base/common/logger";
 import { Iterable } from "src/base/common/util/iterable";
-import { readFileIntoStream, readFileIntoStreamAsync } from "src/base/node/io";
+import { readFileIntoStream, readFileIntoStreamAsync } from "src/code/platform/files/node/io";
 import { createDecorator } from "src/code/common/service/instantiationService/decorator";
 
 export const IFileService = createDecorator<IFileService>('file-service');
