+++ src/code/platform/files/common/storage.ts
@@ -3,7 +3,7 @@ import { DataBuffer } from "src/base/common/file/buffer";
 import { FileOperationError, FileOperationErrorType } from "src/base/common/file/file";
 import { URI } from "src/base/common/file/uri";
 import { ifOrDefault } from "src/base/common/util/type";
-import { IFileService } from "src/code/common/service/fileService/fileService";
+import { IFileService } from "src/code/platform/files/common/fileService";
 
 /**
  * An interface only for {@link DiskStorage}.
