+++ src/code/platform/i18n/i18n.ts
@@ -1,7 +1,7 @@
 import { Emitter, Register } from "src/base/common/event";
 import { URI } from "src/base/common/file/uri";
 import * as path from "src/base/common/file/path";
-import { IFileService } from "src/code/common/service/fileService/fileService";
+import { IFileService } from "src/code/platform/files/common/fileService";
 import { isArray, isObject } from "src/base/common/util/type";
 import { Section } from "src/code/platform/section";
 import { createDecorator } from "src/code/common/service/instantiationService/decorator";
