+++ test/code/platform/i18n.test.ts
@@ -1,6 +1,6 @@
 import * as assert from 'assert';
 import { NullLogger } from 'src/base/common/logger';
-import { FileService, IFileService } from 'src/code/common/service/fileService/fileService';
+import { FileService, IFileService } from 'src/code/platform/files/common/fileService';
 import { i18n } from 'src/code/platform/i18n/i18n';
 import { Section } from 'src/code/platform/section';
